Added LogCourse entity and started progress tracking system (v0.1.18)
Here’s the translation of your text into English: In this commit, I added the LogCourse entity and began developing a system for tracking user progress in courses. A service called LogCourseService was created, which is responsible for interacting with LogCourse in the database. The `progress` method was implemented in LogCourseService, allowing chapters to be marked as completed and returning information about the next chapter. The upsert logic was implemented to update or create a record of the user's progress in the chapter. These changes represent the initial stage of developing the progress tracking system and will be expanded in subsequent commits.